requirement.txt
streamlit
openai
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
python-dotenv
